add survey result test

+++ src/main/java/com/snap/survey/model/response/AnswerChoiceResultResponse.java
@@ -1,3 +1,3 @@
 package com.snap.survey.model.response;
 
-public record AnswerChoiceResultResponse(Integer number, String content, Double percentage) {}
+public record AnswerChoiceResultResponse(Integer number, String content, Integer percentage) {}
